Moved business logic from controllers to services.

+++ src/main/java/com/projekt/services/PriorityServiceImpl.java
@@ -1,5 +1,6 @@
 package com.projekt.services;
 
+import com.projekt.exceptions.*;
 import com.projekt.models.Priority;
 import com.projekt.payload.request.add.AddPriorityRequest;
 import com.projekt.payload.request.update.UpdatePriorityRequest;
@@ -22,7 +23,8 @@ public class PriorityServiceImpl implements PriorityService{
 
     @Override
     public Priority loadById(Long id) {
-        return priorityRepository.getReferenceById(id);
+        return priorityRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Priority", id));
     }
 
     @Override
@@ -31,25 +33,41 @@ public class PriorityServiceImpl implements PriorityService{
     }
 
     @Override
-    public boolean existsByName(String priorityName) {
-        return priorityRepository.existsByNameIgnoreCase(priorityName);
-    }
-
-    @Override
     public void update(UpdatePriorityRequest request) {
-        Priority priority = priorityRepository.getReferenceById(request.priorityID());
+        Priority priority = priorityRepository.findById(request.priorityID())
+                .orElseThrow(() -> new NotFoundException("Priority", request.priorityID()));
+
+        if(priority.getName().equals(request.name())) {
+            throw new NameUnchangedException("Priority", request.name());
+        }
+
+        if(priorityRepository.existsByNameIgnoreCase(request.name())) {
+            throw new NameConflictException("Priority", request.name());
+        }
+
         priority.setName(request.name());
         priority.setMaxTime(request.maxTime());
         priorityRepository.save(priority);
     }
 
     @Override
-    public void save(AddPriorityRequest request) {
+    public void add(AddPriorityRequest request) {
+        if(priorityRepository.existsByNameIgnoreCase(request.name())) {
+            throw new NameConflictException("Priority", request.name());
+        }
+
         priorityRepository.save(new Priority(request.name(), request.maxTime()));
     }
 
     @Override
     public void delete(Long id) {
+        Priority priority = priorityRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Priority", id));
+
+        if(ticketRepository.existsByPriorityId(priority.getId())){
+            throw new ResourceHasAssignedItemsException("priority", "ticket");
+        }
+
         priorityRepository.deleteById(id);
     }
